requirements.txt
pyyaml>=6.0
openai>=1.0.0
google-cloud-bigquery>=3.0.0
